requirements.txt
pillow
streamlit
json
numpy
matplotlib
math
random
pandas